chore: rename tools/fuzz-gen/target/ → subject/ (avoid Rust skip collision)
mache's ingest engine skips directories named 'target/' because
that's the Rust/Cargo build-output convention. mache itself has a
'tools/fuzz-gen/target/' that holds the SUBJECT of the fuzz tester
(the parser being fuzzed) — not Rust build output, but it gets
caught by the same name-based skip and never lands in mache.db.

Concrete impact: ParseToken (defined in target/parser.go and called
from target/parser_test.go's FuzzParseToken) had 0 refs in
node_refs, so dead_code / untested_function flagged it as dead-and-
untested even though the tested-via-call CTE would have caught it
if the file had been ingested.

Rename the directory to 'subject/' (the function being fuzzed is the
subject of the test). Updates:

  • Directory rename via git mv (preserves history).
  • package target → package subject in parser.go and parser_test.go.
  • Taskfile.yml's fuzz-gen task now references subject/parser.go.

Considered alternatives:
  - Removing 'target' from the engine's hardcoded skip list — but
    that's also used by the snapshot copy logic for backing up
    repo state, where Rust target/ exclusion IS desired.
  - Adding contextual detection (Cargo.toml ancestor) — too much
    complexity for a single naming collision in our own tree.

Dogfood: subject/ now indexed (4 nodes, 2 ParseToken refs).
