+++ Taskfile.yml
@@ -104,7 +104,7 @@ tasks:
     desc: Generate a new "Fuzz-Mutate-Repair" task by breaking a target
     dir: tools/fuzz-gen
     cmds:
-      - go run main.go -file target/parser.go -target FuzzParseToken
+      - go run main.go -file subject/parser.go -target FuzzParseToken
 
   check:
     desc: Run all checks (fmt, vet, lint, test, validate)
