+++ tools/fuzz-gen/subject/parser.go
@@ -1,4 +1,4 @@
-package target
+package subject
 
 import (
 	"errors"
