+++ tools/fuzz-gen/subject/parser_test.go
@@ -1,4 +1,4 @@
-package target
+package subject
 
 import (
 	"testing"
